Keep the resumption strategy in table.ts for now
There is currently a proxyquire issue preventing us from moving code out of table.ts so make a TODO and move it later.
